requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.9
django-redis==5.4.0
celery==5.3.4
redis==5.0.1
kafka-python-ng>=2.2.2
ultralytics>=8.2.0
opencv-python==4.9.0.80
easyocr==1.7.1
numpy==1.26.3
scipy==1.11.4
reportlab==4.0.8
Pillow==10.2.0
boto3==1.34.22
django-storages==1.14.2
gunicorn==21.2.0
uvicorn[standard]==0.27.0
daphne==4.1.0
flower==2.0.1
django-celery-beat==2.6.0
python-decouple==3.8
sentry-sdk==1.39.2
whitenoise==6.6.0